requirements.txt
Django==3.1
gunicorn==20.0.4
Pillow==7.2.0
# psycopg2==2.8.5
pytz==2020.1
whitenoise==5.2.0
